requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
WTForms==3.1.2
bcrypt==4.1.3
openai==1.54.3
httpx==0.27.2
gunicorn==21.2.0
python-dotenv==1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
psycopg2-binary==2.9.9

# AI/ML Dependencies for Gemini RAG System
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0

# ChromaDB (Optional - For local development only)
# Uncomment if you want local AI features with vector search
# chromadb==0.4.22